+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/github
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [feacfd6c3]
+- Updated dependencies [83a17db8c]
+  - @runlightyear/lightyear@1.7.0
+
 ## 1.3.0
 
 ### Minor Changes