+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/gmail
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [feacfd6c3]
+- Updated dependencies [83a17db8c]
+  - @runlightyear/lightyear@1.7.0
+
 ## 0.6.1
 
 ### Patch Changes