+++ packages/@runlightyear/gsheets/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/gsheets
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies [feacfd6c3]
+- Updated dependencies [83a17db8c]
+  - @runlightyear/lightyear@1.7.0
+
 ## 0.8.0
 
 ### Minor Changes