+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @runlightyear/postmark
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [feacfd6c3]
+- Updated dependencies [83a17db8c]
+  - @runlightyear/lightyear@1.7.0
+
 ## 0.5.0
 
 ### Minor Changes